+++ resources/generate.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env bash
-
-# Generates a new ISO image to read using xorriso